Delete pkg/ package and split it
