+++ api/client.go
@@ -1,0 +1,11 @@
+package api
+
+import "time"
+
+// ResourceDto represent a resource as given by the API
+type ResourceDto struct {
+	URL   string    `json:"url"`
+	Body  string    `json:"body"`
+	Title string    `json:"title"`
+	Time  time.Time `json:"time"`
+}
