+++ messaging/messaging.go
@@ -1,6 +1,4 @@
-package proto
-
-import "time"
+package messaging
 
 const (
 	// URLTodoSubject is the subject used when an URL is schedule for crawling
@@ -40,12 +38,4 @@ type NewResourceMsg struct {
 // Subject returns the subject where message should be push
 func (msg *NewResourceMsg) Subject() string {
 	return NewResourceSubject
-}
-
-// ResourceDto represent a resource as given by the API
-type ResourceDto struct {
-	URL   string    `json:"url"`
-	Body  string    `json:"body"`
-	Title string    `json:"title"`
-	Time  time.Time `json:"time"`
 }
